Implemented nether block fixes as ABM.
I have tested the change for the min_y but still seems
to be generating invalid terraing due to Engine bug,
so an ABM will trigger when an invalid node is genereted
and will replace with default:stone. I have tested and the
change don't seem to be causing issues with the nether itself.